distributed-slinger/requirements.txt
minimalcluster==0.1.0.dev13
multiprocessing==2.6.2.1
pillow==5.3.0
django==1.11
